Add file cleanup and arrangement scripts
- Implemented `cleanup_old_files.py` to delete files older than 3 months from a specified directory with comprehensive logging and error handling.
- Introduced `file_arranger.py` to organize files into category folders based on file type, supporting recursive scans and dry-run mode.

+++ GeoSentinel-OS/backend/models/fund_usage_model.py
@@ -1,7 +1,8 @@
 """Fund usage records linked to task execution."""
 
+from decimal import Decimal
 from datetime import datetime, timezone
-from sqlalchemy import DateTime, Float, ForeignKey, Integer, String
+from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from database.base import Base
@@ -15,6 +16,6 @@ class FundUsage(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
     task_id: Mapped[int] = mapped_column(ForeignKey("tasks.id"), nullable=False, index=True)
     user_id: Mapped[int | None] = mapped_column(ForeignKey("users.id"), nullable=True, index=True)
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
     description: Mapped[str | None] = mapped_column(String(500), nullable=True)
     spent_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, index=True, default=lambda: datetime.now(timezone.utc))
